requirements.txt
falcon
requests
pycryptodome
redis